moved helper-functions to helper.js, rearrange imp
